+++ src/webhooks/components/WebhookEvents/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./WebhookEvents";
+export * from "./WebhookEvents";