+++ src/webhooks/components/WebhooksList/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./WebhooksList";
+export * from "./WebhooksList";